Move parse method into parse module
